chore: Improve web component published typings
Add index.d.ts instead of .ts files and fix declarations
