+++ jest.config.json
@@ -25,8 +25,8 @@
     "^@templates/(.*)$": "<rootDir>/packages/components/templates/$1",
     "^@pages/(.*)$": "<rootDir>/packages/components/pages/$1",
     "^@components/(.*)$": "<rootDir>/packages/components/$1",
-    "^@types": "<rootDir>/packages/types",
-    "^@hooks/(.*)$": "<rootDir>/packages/hooks/$1"
+    "^@hooks/(.*)$": "<rootDir>/packages/hooks/$1",
+    "^@no-gravity-elements/types": "<rootDir>/packages/types"
   },
   "testMatch": ["**/?(*.)(spec).ts?(x)"],
   "moduleDirectories": ["node_modules"],
