+++ lerna.json
@@ -2,7 +2,7 @@
   "$schema": "node_modules/lerna/schemas/lerna-schema.json",
   "useWorkspaces": true,
   "version": "independent",
-  "packages": ["packages/components/**/*"],
+  "packages": ["packages/**/*"],
   "npmCLient": "yarn",
   "ignoreChanges": [
     "**/*.spec.ts",
