+++ packages/components/molecules/InfoBox/types/index.ts
@@ -1,0 +1,3 @@
+export interface InfoBoxProps {
+  label: string;
+}
