+++ packages/types/index.ts
@@ -1,3 +1,28 @@
+import type { ButtonProps } from '@no-gravity-elements/button/types';
+import type { IconProps } from '@no-gravity-elements/icon/types';
+import type { InfoBoxProps } from '@no-gravity-elements/info-box/types';
+import type { LinkProps } from '@no-gravity-elements/link/types';
+import type { TypographyProps } from '@no-gravity-elements/typography/types';
+
+type ReactWebComponent =
+  | (Partial<HTMLElement> &
+      Partial<React.DetailedHTMLProps<React.HTMLAttributes<HTMLElement>, HTMLElement>>)
+  // @ts-ignore
+  | { children?: React.ReactNode };
+
+declare global {
+  namespace JSX {
+    interface IntrinsicElements {
+      'nge-button': ButtonProps & ReactWebComponent;
+      'nge-card': ReactWebComponent;
+      'nge-icon': IconProps & ReactWebComponent;
+      'nge-link': LinkProps & ReactWebComponent;
+      'nge-typography': TypographyProps & ReactWebComponent;
+      'nge-info-box': InfoBoxProps & ReactWebComponent;
+    }
+  }
+}
+
 export enum TypographySizes {
   h1 = 'h1',
   h2 = 'h2',
@@ -29,4 +54,13 @@ export enum IconNames {
   eye = 'eye',
   menuBurger = 'menu-burger',
   search = 'search',
+}
+
+export enum ButtonTypes {
+  primary = 'primary',
+  secondary = 'secondary',
+  destructive = 'destructive',
+  outline = 'outline',
+  ghost = 'ghost',
+  disabled = 'disabled',
 }
